// apps/api/src/generated/client/enums.ts
/* !!! This is code generated by Prisma. Do not edit directly. !!! */
/* eslint-disable */
// biome-ignore-all lint: generated file
// @ts-nocheck 
/*
* This file exports all enum related types from the schema.
*
* 🟢 You can import this file directly.
*/

export const Role = {
  SUPER_ADMIN: 'SUPER_ADMIN',
  COMPANY_OWNER: 'COMPANY_OWNER',
  HR_ADMIN: 'HR_ADMIN',
  PAYROLL_ADMIN: 'PAYROLL_ADMIN',
  FINANCE_ADMIN: 'FINANCE_ADMIN',
  MANAGER: 'MANAGER',
  EMPLOYEE: 'EMPLOYEE',
  AUDITOR: 'AUDITOR'
} as const

export type Role = (typeof Role)[keyof typeof Role]


export const CountryCode = {
  IN: 'IN',
  US: 'US'
} as const

export type CountryCode = (typeof CountryCode)[keyof typeof CountryCode]


export const EmployeeStatus = {
  ONBOARDING: 'ONBOARDING',
  ACTIVE: 'ACTIVE',
  ON_LEAVE: 'ON_LEAVE',
  TERMINATED: 'TERMINATED'
} as const

export type EmployeeStatus = (typeof EmployeeStatus)[keyof typeof EmployeeStatus]